added exception for failed fit in gscv
