+++ sktime/transformers/detrend/_detrend.py
@@ -6,7 +6,6 @@ __all__ = [
 ]
 __author__ = ["Markus Löning"]
 
-import pandas as pd
 from sklearn.base import clone
 from sktime.forecasting.base._meta import MetaForecasterMixin
 from sktime.transformers.detrend._base import BaseSeriesToSeriesTransformer
